requirements-base.txt
Django==3.2.4
Wand==0.6.6
bleach==3.3.0
boto3==1.17.90
celery==5.1.0
kombu==5.1.0
django-extensions==3.1.3
django-elasticsearch-dsl==7.2.0
git+https://github.com/mytardis/django-form-utils.git@django-1.11-upgrade#egg=django-form-utils
django-jstemplate==1.3.8
django-registration-redux==2.9
django-tastypie==0.14.3
# django-tastypie-swagger
git+https://github.com/mytardis/django-tastypie-swagger.git@mytardis-dj3#egg=django-tastypie-swagger
django-user-agents==0.4.0
django-webpack-loader==1.0.0
django-widget-tweaks==1.4.8
django-storages==1.11.1
gunicorn==20.1.0
gevent==21.1.2
html2text==2020.1.16
html5lib==1.1
isodate==0.6.0
pyjwt==2.1.0
lxml==4.6.3
paramiko==2.7.2
pillow==8.2.0
pyoai==2.5.0
pystache==0.5.4
python-dateutil==2.8.1
# python-magic
git+https://github.com/mytardis/python-magic.git#egg=python-magic
pytz==2021.1
requests==2.25.1
uritemplate==3.0.1
# GraphQL
graphene-django==2.8.0
graphene-django-plus==1.3.2
django-filter==2.2.0
django-crispy-forms==1.8.1
django-graphql-jwt==0.3.1
django-graphiql==0.4.4
django-cors-headers==3.2.1
